// frontend/src/app/learncoreservices/models/user-profile-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UserProfileResponse {
  avatarUrl?: string;
  createdDate?: string;
  dateOfBirth?: string;
  email?: string;
  firstname?: string;
  lastname?: string;
  roles?: Array<string>;
  tenantUuid?: string;
  username?: string;
  uuid?: string;
}